Refactor (create a UniqueEntryBaseModel)
