+++ src/base/models.py
@@ -1,3 +1,4 @@
+from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
@@ -29,3 +30,18 @@ class DurationBaseModel(models.Model):
         start_date = self.start_date.strftime('%b %Y')
         end_date = self.end_date.strftime('%b %Y') if self.end_date else _('Present')
         return f"{start_date} - {end_date}"
+
+
+class UniqueEntryBaseModel(models.Model):
+    class Meta:
+        abstract = True
+
+    def save(self, *args, **kwargs):
+        # Ensures that there is only one entry in the table
+        model_class = self.__class__
+        if model_class.objects.exists() and not self.pk:
+            raise ValidationError(_(f"There can only be one entry in the {model_class.__name__} table."))
+        super().save(*args, **kwargs)
+
+    def delete(self, *args, **kwargs):
+        raise ValidationError(_(f"Deleting the {self.__class__.__name__} object is prohibited."))
